refactor: channels details table and dialog into component
