+++ src/app/core/components/channel-details-dialog/channel-details-dialog.component.scss
