+++ src/app/core/components/channel-details-dialog/mail-channel-details/mail-channel-details.component.ts
