+++ src/app/core/components/channel-details-dialog/phone-channel-details/phone-channel-details.component.html
