+++ src/app/core/components/channel-details-dialog/phone-channel-details/phone-channel-details.component.ts
@@ -1,4 +1,4 @@
-import { ChangeDetectorRef, Component, OnDestroy, OnInit, forwardRef } from '@angular/core';
+import { Component, OnDestroy, forwardRef } from '@angular/core';
 import { AbstractControl, FormBuilder, NG_VALIDATORS, NG_VALUE_ACCESSOR, ValidationErrors, Validator, Validators } from '@angular/forms';
 import { Subscription } from 'rxjs';
 import { PhoneChannelDetails } from 'src/app/core/model/channel';
