+++ src/app/core/components/channel-details-dialog/radio-channel-details/radio-channel-details.component.html
