+++ src/app/features/logistics/logistics.module.ts
@@ -1,30 +1,23 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { LogisticsLayoutComponent } from './logistics-layout/logistics-layout.component';
-import { AddressBookEntryListView } from './address-book-entry-list-view/address-book-entry-list-view.component';
+import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
-import { RouterLinkWithHref, RouterOutlet } from '@angular/router';
 import { MatNativeDateModule } from '@angular/material/core';
-import { MtxDatetimepickerModule } from '@ng-matero/extensions/datetimepicker';
+import { RouterLinkWithHref, RouterOutlet } from '@angular/router';
 import { MtxNativeDatetimeModule } from '@ng-matero/extensions/core';
+import { MtxDatetimepickerModule } from '@ng-matero/extensions/datetimepicker';
+import { CoreModule } from '../../core/core.module';
+import { AngularMaterialModule } from '../../core/util/angular-material.module';
+import { AddressBookEntryAutoIntelDeliveryComponent } from './address-book-entry-list-view/address-book-entry-auto-intel-delivery/address-book-entry-auto-intel-delivery.component';
+import { AddressBookEntryListView } from './address-book-entry-list-view/address-book-entry-list-view.component';
 import {
   CreateAddressBookEntryView,
 } from './address-book-entry-list-view/create-address-book-entry-view/create-address-book-entry-view.component';
 import {
   EditAddressBookEntryView,
 } from './address-book-entry-list-view/edit-address-book-entry-view/edit-address-book-entry-view.component';
-import { ChannelsComponent } from './channels/channels.component';
-import { CoreModule } from '../../core/core.module';
-import { AngularMaterialModule } from '../../core/util/angular-material.module';
-import { ChannelDetailsDialog } from './channels/channel-details-dialog/channel-details-dialog.component';
-import {
-  RadioChannelDetailsComponent,
-} from './channels/channel-details-dialog/radio-channel-details/radio-channel-details.component';
-import { IntelDeliveryModule } from './intel-delivery/intel-delivery.module';
-import { AddressBookEntryAutoIntelDeliveryComponent } from './address-book-entry-list-view/address-book-entry-auto-intel-delivery/address-book-entry-auto-intel-delivery.component';
 import { AddressBookLayoutComponent } from './address-book-layout/address-book-layout.component';
-import { MailChannelDetailsComponent } from './channels/channel-details-dialog/mail-channel-details/mail-channel-details.component';
-import { PhoneChannelDetailsComponent } from './channels/channel-details-dialog/phone-channel-details/phone-channel-details.component';
+import { IntelDeliveryModule } from './intel-delivery/intel-delivery.module';
+import { LogisticsLayoutComponent } from './logistics-layout/logistics-layout.component';
 
 
 @NgModule({
@@ -34,13 +27,8 @@ import { PhoneChannelDetailsComponent } from './channels/channel-details-dialog/
     CreateAddressBookEntryView,
     EditAddressBookEntryView,
     LogisticsLayoutComponent,
-    ChannelsComponent,
-    ChannelDetailsDialog,
-    RadioChannelDetailsComponent,
     AddressBookEntryAutoIntelDeliveryComponent,
-    AddressBookLayoutComponent,
-    MailChannelDetailsComponent,
-    PhoneChannelDetailsComponent
+    AddressBookLayoutComponent
   ],
   imports: [
     CommonModule,
